+++ tests/integration/database/midgard/_Enums.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace mako\tests\integration\database\midgard;
+
+enum FooEnum: int
+{
+	case ONE = 1;
+	case TWO = 2;
+}